+++ tests/graph_validation_test_runner/translator/trapi/test_trapi.py
@@ -8,11 +8,6 @@ from graph_validation_tests.translator.trapi import (
     get_component_infores_object_id,
     resolve_component_endpoint
 )
-from graph_validation_tests.utils.http import post_query
-from graph_validation_tests.utils.ontology_kp import (
-    ONTOLOGY_KP_TRAPI_SERVER,
-    NODE_NORMALIZER_SERVER
-)
 
 pytest_plugins = ('pytest_asyncio',)
 
@@ -66,72 +61,6 @@ def test_resolve_component_endpoint(
             target_biolink_version=None
         )
     assert endpoint == result
-
-
-@pytest.mark.parametrize(
-    "curie,category,result",
-    [
-        (   # Query 0 - chemical compounds are NOT in ontology hierarchy
-            "CHEMBL.COMPOUND:CHEMBL2333026",
-            "biolink:SmallMolecule",
-            None
-        ),
-        (   # Query 1 - MONDO disease terms are in an ontology term hierarchy
-            "MONDO:0011027",
-            "biolink:Disease",
-            "MONDO:0015967"
-        )
-    ]
-)
-@pytest.mark.asyncio
-async def test_post_query_to_ontology_kp(curie: str, category: str, result: Optional[str]):
-    query = {
-        "message": {
-            "query_graph": {
-                "nodes": {
-                    "a": {
-                        "ids": [curie]
-                    },
-                    "b": {
-                        "categories": [category]
-                    }
-                },
-                "edges": {
-                    "ab": {
-                        "subject": "a",
-                        "object": "b",
-                        "predicates": ["biolink:subclass_of"]
-                    }
-                }
-            }
-        }
-    }
-    response = post_query(url=ONTOLOGY_KP_TRAPI_SERVER, query=query, server="Post Ontology KP Query")
-    assert response
-
-
-@pytest.mark.parametrize(
-    "curie,category",
-    [
-        # Query 0 - HGNC id
-        ("HGNC:12791", "biolink:Gene"),
-
-        # Query 1 - MONDO term
-        ("MONDO:0011027", "biolink:Disease"),
-
-        # Query 2 - HP term
-        ("HP:0040068", "biolink:PhenotypicFeature")
-    ]
-)
-@pytest.mark.asyncio
-async def test_post_query_to_node_normalization(curie: str, category: str):
-    j = {'curies': [curie]}
-    result = post_query(url=NODE_NORMALIZER_SERVER, query=j, server="Post Node Normalizer Query")
-    assert result
-    assert curie in result
-    assert "equivalent_identifiers" in result[curie]
-    assert len(result[curie]["equivalent_identifiers"])
-    assert category in result[curie]["type"]
 
 
 # @pytest.mark.asyncio