+++ tests/standards_validation_test_runner/test_standards_validation_test.py
@@ -2,13 +2,23 @@
 Unit tests for Standards Validation Test code validation
 """
 from sys import stderr
-from typing import Dict
+from typing import List, Dict
 from json import dump
 import pytest
 
-from graph_validation_tests.utils.unit_test_templates import by_subject, by_object
-from standards_validation_test_runner import StandardsValidationTest, run_standards_validation_tests
-from tests import SAMPLE_MOLEPRO_INPUT_DATA, SAMPLE_ARAX_INPUT_DATA, SAMPLE_ARAGORN_INPUT_DATA
+from graph_validation_tests.utils.unit_test_templates import (
+    by_subject,
+    by_object
+)
+from standards_validation_test_runner import (
+    StandardsValidationTest,
+    run_standards_validation_tests
+)
+from tests import (
+    SAMPLE_MOLEPRO_INPUT_DATA,
+    SAMPLE_ARAX_INPUT_DATA,
+    SAMPLE_ARAGORN_INPUT_DATA
+)
 
 
 @pytest.mark.parametrize(
@@ -71,24 +81,30 @@ async def test_standards_validation_test_on_ars():
 
 
 @pytest.mark.parametrize(
-    "data,environment,component,expected_result",
+    "data,environment,components,expected_result",
     [
-        (
+        (   # Query 0 - MolePro test
             SAMPLE_MOLEPRO_INPUT_DATA,
             "ci",
-            "molepro",
+            ["molepro"],
             None
         ),
-        (
+        (   # Query 1 - ARAX test
             SAMPLE_ARAX_INPUT_DATA,
             "ci",
-            "arax",
+            ["arax"],
             None
         ),
-        (
+        (   # Query 2 - Aragorn test
             SAMPLE_ARAGORN_INPUT_DATA,
             "ci",
-            "aragorn",
+            ["aragorn"],
+            None
+        ),
+        (   # Query 3 - Combined ARAX and MolePro test
+            SAMPLE_MOLEPRO_INPUT_DATA,
+            "ci",
+            ["molepro", "arax", "aragorn"],
             None
         )
     ]
@@ -97,13 +113,13 @@ async def test_standards_validation_test_on_ars():
 async def test_run_standards_validation_tests(
         data: Dict,
         environment: str,
-        component: str,
+        components: List[str],
         expected_result
 ):
     results: Dict = await run_standards_validation_tests(
         **data,
         environment=environment,
-        components=[component]
+        components=components
     )
     assert results
     dump(results, stderr, indent=4)