+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+# 0.1.4
+
+- upgrade to reasoner-validator 4.2.0
+- fix bug in completely returning of result message dictionaries from test runs with a list of more than one target component
+
 ## 0.1.3
 
 - made ARS unimplemented TRAPI query non-fatal with error messages